core/requirements.txt
requests>=2.31.0
pyecharts>=2.0.0
python-dotenv>=1.0.0
psutil>=5.9.0
pycaw>=20230407
comtypes>=1.2.0
lark-oapi>=1.4.0
